Detectar colisiones entre cadenas

+++ src/main/java/org/brotherhood/mutantdna/entities/Stats.java
@@ -25,7 +25,7 @@ public class Stats {
 
 	@JsonProperty("ratio")
 	public double getRatio() {
-		return countMutantDna/countHumanDna;
+		return (double)countMutantDna/(double)countHumanDna;
 	}
 
 	@Override
